requirements.txt
alembic==1.8.1
APScheduler==3.9.1.post1
attrs==22.1.0
bcrypt==4.0.1
bidict==0.22.0
bleach==5.0.1
blinker==1.5
Brotli==1.0.9
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
configparser==5.3.0
cryptography==38.0.4
cyclonedx-python-lib==3.1.1
DateTime==4.7
dnspython==2.2.1
email-validator==1.3.0
exceptiongroup==1.0.4
Flask==2.2.2
Flask-Compress==1.13
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-mailman==0.3.0
Flask-Migrate==2.7.0
Flask-Moment==1.0.5
Flask-Principal==0.4.0
Flask-Script==2.0.5
Flask-Security-Too==5.0.2
Flask-SocketIO==5.3.2
Flask-SQLAlchemy==2.5.1
flask-talisman==1.0.0
Flask-WTF==1.0.1
greenlet==2.0.1
gunicorn==20.1.0
h11==0.14.0
html5lib==1.1
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
lockfile==0.12.2
Mako==1.2.4
MarkupSafe==2.1.1
mkdocs-material-extensions==1.1.1
more-itertools==9.0.0
msgpack==1.0.4
mysqlclient==2.1.1
packageurl-python==0.10.4
packaging==21.3
passlib==1.7.4
pip-api==0.0.30
pip-autoremove==0.10.0
pip-requirements-parser==32.0.0
pip_audit==2.4.8
pluggy==1.0.0
pycparser==2.21
Pygments==2.13.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pytest==7.2.0
python-engineio==4.3.4
python-http-client==3.3.7
python-socketio==5.7.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
resolvelib==0.9.0
rich==12.6.0
sendgrid==5.6.0
simple-websocket==0.9.0
simplejson==3.18.0
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.45
SQLAlchemy-Utils==0.38.3
toml==0.10.2
tomli==2.0.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
webencodings==0.5.1
Werkzeug==2.2.2
wsproto==1.2.0
WTForms==3.0.1
zope.interface==5.5.2
